Fix type annotation

+++ regression/python/function-call-mult-params/main.py
@@ -1,4 +1,4 @@
 def func(a : int, b: int) -> None:
-    result : int = a/b
+    result:float = a/b
 
 func(1,0)
